rename decode() to lower(), move into vm module
