+++ src/util.rs
@@ -1,0 +1,6 @@
+#![macro_use]
+
+#[macro_export]
+macro_rules! trace(
+    ( $($thing:expr),* ) => { if cfg!(test) { println! { $($thing),* } } };
+);
